Refactor ActionControllerPatch to be more ruby idiomatic
